build: improved entry modules
The private packages have their "main" set to "src/index.ts" or similar. Many of the public packages transform "src/index.ts" into "index.cjs.js". The `react-spring` and `@react-spring/shared` packages are the exceptions.

This makes it easier for bundlers (like Metro) to watch for changes while maintainers are developing react-spring with a "yarn link" workflow.

Also, the bundles are no longer put in the "src" directory alongside the source files. This makes it easy to distinguish the two from each other.