fix: look for "animation.config" before entering the FrameLoop
in both "SpringValue#_start" and "SpringValue#_stop"

TLDR: Keep "To" springs out of the frameloop, but not their children!

This is important because child "To" springs should never enter the frameloop, but the children of those "To" springs *should* have a chance to enter the frameloop if they want to animate *toward* the value of a "To" spring.