fix: "onRest" calls in Controller
Previously, the promise returned from "Controller#start" was used to track when "onRest" should be called. There are multiple problems with that; the main one being the "delay" prop causing a crash via unexpected internal state.

The mental model has been simplified by keeping a map of spring values to their last animation result. This is much easier to reason about, because it relies on when SpringValue calls its "onRest" prop, instead of when the "start" promises are resolved.