feat: add "onProps" prop + remove "onAnimate" prop
The "onAnimate" prop was called before props were merged into the animation.

The "onProps" prop is called *after* props are merged. This is better, because this animation will have the newest goal value, etc. And I didn't want to encourage prop manipulation in "onAnimate" (too much action at a distance).