fix: have "onRest" handlers determine their own "finished" value
This fixes an edge case where an update with an async "to" prop also contains "reverse: true" which could be interrupted by an async update, and its "finished" value would be false even if the current value equals the goal value.
There are probably some other edge cases fixed as well.