fix: avoid merging other props when "to" is temporally prevented
All other props (except "from") must not be merged if the "to" prop
was temporally prevented by an earlier update. If the "to" prop is
undefined, the delayed update cannot be temporally prevented.

For the "from" prop to be temporally prevented, an earlier update
needs to set the "from" prop.

It should be noted that this is on a per-key basis. So if a delayed
update has { a: 1, b: 1 } as its "to" prop, and an earlier update
has only { a: 0 } as its "to" prop, then the { b: 1 } part will
still be used.