test: fix how animations are observed
Since delayed updates are synchronized with the frameloop now, we can't simply access the internal "FrameLoop#_animations" array. Instead, we need to inject a special handler that the frameloop calls immediately before it advances the active animations.

Also, expose the "FrameLoop#_dispose" method for better disposal of the frameloop after each test.