feat(types): re-export the CSS namespace from csstype
This is useful when you want to update a style prop like `position` (which expects a `CSS.PositionProperty` string type) using the react-spring API. Users must type-cast the position value (eg: "absolute" as CSS.PositionProperty) for its inferred type to be compatible with the `style` prop of a React element. It will be easier for users to import `CSS` from react-spring instead of installing "csstype" separately.